Not working implementation of Powerbuilder button.  See below.
For reasons I'm too Python-newb to understand, inside the constructor of
my PowerBuilderButton, it's the right class.  As soon as it gets out of
there, it's just a wx.Button.  This won't do.  Would rather figure out
why it's broken than fall back to making each of these buttons by hand
with tiny magnetic needles touching the hard drive.