Make ControlGroup controls auto-Enable() their labels
This does some dances with mixins and subclasses so that calling
.Enable() on a control instantiated by ControlGroup will automatically
call it also on the CtlLabel associated with that control.

There are probably a few places where we need to use the new cg* classes
in places we create controls by hand, but we'll find those as we go, I
suppose.