Rework all """ strings to try to avoid malware false positives
This release tinkers a few places where informational strings were
built, using a different scheme that some suggest might not trigger
malware false positives.  We won't actually know until this builds and
can be tested.

This release is functionally exactly the same as the previous one, so if
that one is working for you, you don't need to rush to update.