Add "Buff Display" command to PowerBinder.
This release adds a new command to PowerBinder that allows creating
binds to set the "Buffsettings" game option to control the display of
buffs in the status, group, and pets windows.  More details here:

https://homecoming.wiki/wiki/Buff_Icon_Display_Options

It also fixes a bug where if a Profile were renamed, Ctrl-S would save
it to the previous filename.  It will now correctly save to a profile
named for the current "Name" entry in the General tab.