Add various Inherent and SMART vendor powers to the PowerPicker
This release updates the PowerPicker popup to contain various inherent,
SMART vendor, and accolade-based powers that weren't available in
BindControl any other place.  There are probably still powers in the
game that are missing from the picker, mostly a giant list of temporary
jetpacks and the like, but this covers lots of powers people might want
permanent binds for.