Global check key conflicts on load and every change.
This release much more aggressively checks the whole set of key binds
looking for overlapping/conflicting definitions.  It will do this at
load time, as well as any time a keybind changes, and will highlight key
buttons that are in conflict, throughout the app.  There's probably
still some behavior weirdness and some cases where it should check that
it currently isn't, but this is a feature I've wanted for quite a while.

There will be more work on the UI of this.