Add tray button binds to Gameplay tab
This release adds the option to re-bind the keys that trigger the
various tray slots in the default UI.  The defaults are the same as the
"Modern" scheme from Homecoming Issue 27, Page 7.

If you aren't going to change up the game's default power-tray keybinds,
you can leave this whole section disabled/unchecked.  This is just for
people who like quite a large degree of control over their binds.