"The scrollbar release"
This release normalizes and clarifies a lot of the layout issues that
were causing various kinda of scrollbar chaos, including scrollbars when
none were needed, and no scrollbars when one was needed.

Each tab manages its own scrollbars and should only display them when
the app is too small for that specific tab's contents.  This gets rid of
the ocean of scrollable vertical space at the bottom of shorter tabs,
and has the nice side effect of "pinning" the tabs at the top of the
screen at all times.

Please file issues if there's any weirdness with layout and scrolling.