Better handling of default profile
It's been possible to get in a state where you're working directly with
the default profile, which is not intended.  This release does some due
diligence to try to keep the default profile from being the active one
in the editor.  There's probably still more work to be done on this
front.