All-new handling of default profile.
This release again reworks the default profile handling, this time to an
all-new system.

-- The default profile is now saved with the application preferences,
and is not directly editable.
-- The profile file "Default.bcp" is no longer treated specially in any
way except as noted just below.
-- The first time you do a "New Profile", your old Default.bcp, if any,
will be copied to the new location if nothing is already saved there.
Default.bcp will not be automatically deleted.

To avoid confusion, if you have a Default.bcp, you're encouraged to load
it, "Save As Default Profile," and delete it.  There's no harm in not
doing this, but "Default.bcp" is no longer magical in any way, so
changes to it will not affect "New Profile" going forward.

This involved some surgery to loading and saving of profiles in general,
and while I'm pretty confident in my testing of various cases, it's
possible I missed something.  Please open issues as you find them.
Profile contents have not changed, so it's safe to revert to a previous
version if this one gives you troubles.