Actual release of "Delete All Binds" functionality.
This release adds a new "Delete All Binds" button to the bottom, next to
"Write Binds."

This will attempt to delete all BindControl-managed files in the current
Profile's binds directory and remove empty directories (including the
binds directory itself). Manually-added files will not be touched.

It feels pretty solid, but programmatically deleting files is always
nervous-making. The feature attempts to make sure you don't have your
binds directory unset or set to the root of a drive or something; it
also will let you know what path it's about to go spelunking in and asks
for confirmation.

Please check it out if this functionality interests you. Please be
careful with it and let me know if it does anything remotely surprising,
by filing an issue. Thanks.