v0.21.0.1 - More robust generation of binds directory candidates
This release does a more thorough job of trying to generate a unique profile-specific binds directory.  It also checks better for existing profiles' binds directories and will try not to pick those.

The actual upshot is that it's slightly more likely just to throw its hands up and pop up the dialog box asking you for a profile binds dir.