Add "Chatty is default" preference to Mastermind Page
This release adds a "chatty is default" checkbox to the Mastermind page,
which will select whether pets reset to chatty or non-chatty mode when
the Binds Reset key is pressed.

It also fixes a misbehavior bug when "Load Last Profile" is selected as
a preference and the last profile can't be found, ie, has been deleted
or moved.