Reorganization of Movement / Speed on Demand tab code.
Apart from a few tiny display and language twiddles, the MovementPowers tab should act just like the previous release, except that it should be much much less laggy and jumpy when changing any controls that change the state of it, ie, that enable or disable or hide or show other controls.

This is mostly an internal code organization release, but it enables further upcoming work on the Movement Powers tab to be easier and safer.