Beta-quality support for Rebirth.  Read on:
This release adds support for Rebirth to BindControl.

Specifically, updated functionality includes:
* Correct and updated Archetype / powerset data, including Guardian
* Changes / additions to Power Pools
* Support for the "Genesis" Incarnate Slot
* Updated emotes and other smaller changes

Most of the data changes were gleaned from long hours squinting at Mids Reborn, as well as perusing the Update Notes on the Rebirth wiki.

Known TODOs:

* Teleport powers currently use the Homecoming-specific /powexec_location slash command.  The Movement Powers page in general has a lot of Homecoming-think that needs revisiting.
* Rebirth's abbreviated slash commands are not supported yet, making binds longer than they need to be.
* The Popmenu Editor only works with Homecoming installations.

These are all very high on the list.  Rebirth players are asked to try it out and let me know where all I got it wrong.  I'm just getting started on Rebirth so I don't have high level toons to test with etc, so bear with me as we iron out the wrinkles.

It's **certain** that some things were missed and/or are just wrong.

Find a bug?  Feedback on how I could make it better?  File an issue!

This is a big change.  There will be bugs.  Some of them may even crash the program.  Let's make it better.

<hr>

Oh, before you ask -- new profiles default to Homecoming.  If you want to make it default to Rebirth instead, make a new Profile, switch it to Rebirth, then "Save as Default Profile" and voila.