Major rework of Profile creation and loading to try to reduce overhead
This release is an internal reorganization of the code that creates Profile objects, both via "New Profile" and via loading Profiles from file, either explicitly or at startup.  Because of streamlining and process gains from this, the app starts up over a half second faster on Windows now.  Hopefully more work on these lines can be done.

This release also includes Rebirth support in the Popmenu Editor, via understanding Rebirth installations when setting the GameDir.  There is still some work to be done here, as BindControl still has only one notion of a GameDir, meaning the Popmenu Editor doesn't honor the Profile's notion of which Server it's using.  This is on the list to fix.