New "Bind Wizard" functionality, with "Incarnate Set" wizard included
This release creates a new button on the Custom Binds page, "Launch
Custom Bind Wizard" that will help the user create more elaborate and
complicated binds that might be too fiddly to do by hand using a Complex
Bind.

Several Bind Wizards are planned, but this release ships just the first
one, "Incarnate Set," which allows the user to set up a full or partial
set of Incarnate Powers that can then be slotted using multiple presses
of the bound key.

There are also a number of wee bugfixes and functionality updates, as
well as almost certainly some new bugs related to the Wizard
functionality and elsewhere.

Please check it out and file Github issues as you find things.