PowerBinder additions and updates
This release adds a few new slash commands to PowerBinder, mostly having
to do with camera control, though also a few graphics settings.

It also reworks some existing PowerBinder commands to have tooltips, and
to lay themselves out better inside the PowerBinder dialog.

Finally, this release adds SlashCommands.md at the root of the repo,
which is basically a checklist / work-in-progress / TODO list showing
the slash commands that still either need to be implemented or rejected
for implementation.  There's a lot of them left but many many of them
simply don't make sense as keybinds and will end up rejected.