Custom Binds' bindstrings are now read-only in the main UI
This has been a long-desired change (by me, at least) -- until now, it
was possible to edit the bindstring in a Simple Bind or in a Complex
Bind's step directly, which would make it fall out of sync with the
state of the PowerBinder configuration that, in principle, was
responsible for managing that bindstring.

Some people had been using that to create completely custom
bindstrings, but then would discover that if they opened the PowerBinder
editor, it would be easy to splat over their changes accidentally.

Starting with this release, the bindstrings are read-only and cannot be
changed directly.  The intended way to create a custom command in your
bindstring is by using the "Custom Command" entry from the PowerBinder
"Add Command" picker.

As part of this mild overhaul of PowerBinder, the PowerBinder dialog's
list of entries has been rebranded from "Steps" to "Commands" to clear
up some linguistic confusion with Complex Binds' "Steps" which are a
different thing.

This will hopefully not break anything for anyone too terribly.  If you
find problems, bugs, or feature requests, please open a Github issue
about them.