Lazy-create the Wizard Dialog:
This fixes the bug on Windows where the dialog was being created at
start-time, if there was a Wizard BindPane, with a pathological size and
location, making it not appear on edit of that bindpane.  This was a big
evil tangle and probably has a bug or two in it still, but I think it's
working better than not being able to edit Incarnate Set Binds from a
loaded Profile.