Add Temporary Travel Powers to Movement Powers Page
This has been a long time coming.  CityBinder, back in the day, had a
notion of triggering Temp Travel Powers, but relied on the power being
in a particular tray and slot.  It also tried to integrate the temp
powers into the Speed on Demand framework, which sometimes had hilarious
results like turning off your Jet Pack every time you stopped moving.

BindControl now offers a simple toggle keybind for any Temp Travel
Power, selected by name.  It's super basic, and is essentially a Simple
Bind hiding on the Movement Powers Page.

There was some underlying work done on the PowerPicker in general, which
has actually cleaned up and clarified that functionality in several
helpful ways.  If you notice any weirdness using it with PowerBinder to
make Simple or Complex Binds, please file an issue, but it should Just
Work.