Big rework of Buffer Binds.
This release features a long-overdue rework of the Buffer Bind pane,
making it a bit clearer what's going on, removing the confusing
PowerBinder buttons, and cleaning up some of the internal logic.

There will probably be further work in the near future, making it so
that Buffer Binds aren't hard-coded to having three buffs, which is at
the very list sort of UI-confusing if you only want one or two.

This seems to work well for me, but might have some bugs.  Please open
issues as you find things.

v0.28.2.1 bugfix:  The app will no longer pester with an intrusive
dialog if the Gamedir is not correctly set up, at least not at startup.