Phase II of major rework of Buffer Binds
With this release, Buffer Binds now start with one buff, and the user
can add (or delete) buffs as they see fit, instead of having a
hard-coded three of them sitting around, often needing to be skipped
over in code if there was nothing there.  This has clarified and
simplified some of the innards of Buffer Binds, as well as cleaned up
some of the display code.

This involved changing how Buffer Binds are saved in the Profile.  Old
profiles with the "old way" will still load correctly, and will save as
the "new way" the next time they're saved.

A couple of other wee tweaks sneaked in, including that the Power Picker
buttons no longer have that sorta misleading (and frustrating, in code)
"..." when they're not populated -- they're just blank, and typically
will be in an error condition, though not currently in all cases.

As always, please open issues as you find them.