Separate Profile initialization from population
This is an internal reorganization of how the Profile data structure and
UI get set up.

This is in pursuit of some other projects I'm working on that wanted to
have the Profile completely initialized before it started in on
populating the UI.  This is a long-desired change, and is just a couple
of steps away from getting the Profile completely separate from its UI,
which would allow for a lot more by way of automated testing.

Anyway, it should act just like it did before, except, as a side effect,
now brand-new profiles will be pre-flagged as "modified" so will prompt
you to save them in all the appropriate circumstances.  If you find
anything else that's changed or broken, please open an issue.