Complex Binds:  warn more correctly about steps > 255 chars
Complex Binds add a longish BLF() to the end of each step in order to
bump it along to the next step.  The UI was not taking this additional
length into account when warning about steps being too long, though it
did correctly (and therefore surprisingly) warn and not-write during the
"Write Binds" step after the BLF() was added.  This commit fixes that.