Big Mastermind Update
This release is yet another big revamp of the Mastermind page.  In
particular, two new bind schemes have been added, derived from posts to
the Homecoming forums by qwy in 2020.  These are intended as
alternatives to the Sandolphan binds, and only one of the three can be
active per Profile.

Information about the new bind schemes can be found in the little blue
[?] help buttons on the appropriate tabs on the Mastermind page.

Specific release notes:
* Added qwyNumpad and qwyPetMouse bind schemes
* Reorganize+refine pet-names / rename / MM level controls
* Moved "select next" and "select prev" binds into the Sandolphan scheme
* Numpad bugfixes with the KeySelect dialog
* Updates to Popmenu Editor to accommodate qwyPetMouse optional popmenu
* Many text, help, and tooltip updates to explain all of this
* Many wee bugfixes and reorganizations of code
* New unit tests for various bits of this