PowerBinder:  return CurrentState on SaveToData instead of making dialog
This is an attempt to fix a bug that causes long saving times on
Profiles with many PowerBinders in them.

This is an experimental release -- just use v0.31 if you are not having
trouble with long save times.

This is only trivially tested, so back up your Profiles before saving
them if you use any PowerBinders at all.