Github Actions update to try to automate prerelease / test handling
This is a packaging-only release, experimenting with automating whether
a release is a prerelease and whether it is flagged as the latest
release based solely on the format of the tag.

There might be a few releases in here as I iterate this.  Apologies.

Continue to use v0.31 for now.