Make PowerBinderDialog only on-demand;  keep state in PowerBinder
This test release does further work to minimize the impact of having
many PowerBinders in a Profile.  It had been possible to fall afoul of
Windows making too many entities because we were making edit dialogs at
the drop of a hat, possibly having one edit dialog for each PowerBinder
sitting around ready to be displayed, chock full of buttons and lists
and controls that counted against Windows' entity limit.

With this change, dialogs are instantiated as needed and should be
cleaned up and garbage-collected once they're closed.  This seems to
work for me, so let's try it out.