Fix CustomID versus duplicating binds problem, retroactively too
Duplicating a Custom Bind would, in basically every case, cause a
CustomID collision between the original bind and the duplicate one.
This would be very bad.

This bugfix corrects the problem, AND detects any profiles that were
saved that way and auto-heals them to have unique CustomIDs when the
Profile is loaded and re-saved.