First test release of the Big Testing Rework
This release is the largest and most comprehensive change ever made to
BindControl.  It is for the brave and/or foolhardy -- if you just want
to keep tinkering with your keybinds undisturbed, stick with v0.31.

In the name of making the code testable, this contains a complete rework
of the way that the data representing the Profile are held in memory,
loaded, saved, and validated.

I also took the opportunity to pick through the code with the assistance
of various linter and verifier tools, which revealed many many small
instances of bad practice, as well as several actual bugs.

Many more tiny changes have happened, including bugfixes, small feature
updates, and documentation additions and clarifications.

Despite all of these deep changes, BindControl should for the most part
look and act exactly like it has up until now -- in fact, any
behavior change is probably a bug.

And there will be bugs.  In a change of this size, things will have been
missed or incorrectly implemented or just broken.  Please try out every
corner of BindControl that you feel up to testing, and report bugs as
Github issues or in the Homecoming Forums thread.

This change will help avoid regressions and new bugs, going forward,
especially as more and more tests are written, so it's going to be a
very good thing for the future of BindControl, but it's going to be a
little bit of a slog getting there.