Test release:  "Escape Configurator" Wizard
This is a test release of an experimental "Escape Configurator" Bind
Wizard.  It's pretty simple, allowing the user to select any or all of
several different commands to run when ESC is pressed.  It's nothing
that couldn't be done with just a SimpleBind, but this consolidates the
config into one easy-to-use place.

Currently, as it uses the same framework as the Incarnate Set wizard, it
has its contents show up on the bind page, but you click on that to
launch a dialog to configure it.  That's fine, but I may experiment with
using the contents on the bind page as the configurator, skipping the
separate dialog.

All feedback welcome, in the Github issues page, or in the Homecoming
forums thread.