Escape Configurator;  bugfixes
This is the actual release of the "Escape Configurator" Bind Wizard.  It
might still want some cleanup and UI love but it seems to work well
enough for release.  With this Bind Wizard, you can configure the Escape
key to do any number of possibly-useful things.  Some combinations of
things are probably less useful and more annoying than others, but
they're all in there to be experimented with.

Also in here is further work on the cleanup of the internal
representation of Profile Data, which has been simplified even further,
and a couple of bugfixes.

Finally, the basic no-filename versions of /wdwsave and /wdwload have
been added to PowerBinder.