EXTREMELY experimental and untested "edit default profile" functionality
With this experimental test release, a new menu item has appeared,
Profile > Edit Default Profile.  This will go through the normal
profile-closing steps if needed, and then re-open BindControl in a
special mode containing the default profile.

In this mode, the interface is an ugly light blue color, and a warning
is thrown at the bottom to indicate that you are editing the default
profile and not an actual one.

The normal "Save Profile" menu item and the "Write Binds" etc buttons at
the bottom are disabled.  The only save option is "Save as Default
Profile" in the Profile menu, as usual.

Closing the profile / loading a new profile / closing the app, when
changes have been made, will offer the normal "DEFAULT PROFILE has
changed, save changes?" warning, and will save to the default profile
slot if you answer yes.

This is extremely ugly and not rigorously tested.  It's likely y'all
might be able to find a way to make it save the default profile as a
normal one, or otherwise break the intended flow.  File bugs if you do.
Give feedback on the UI changes.  I know the color is ugly, but you
should have seen some of the other ones I tried.

Anyway, this is just a WIP I'm throwing out there to get feedback, and
it might or might not show up in anything resembling this form.  Let me
know what you think.