BUGFIX:  Changing ProfileBindsDir was not getting saved with Profile
This commit removes the whole notion of Profile.ProfileBindsDir as an
attribute, instead making it an accessor that points to
Profile.Data['ProfileBindsDir'] which is where we actually keep it.

It also makes good and sure that the latter is actually set in every
possible case, except the ones I've missed somehow.  Anyway it should be
a bit more bullet-resistant to ancient Profiles that don't have a
ProfileBindsDir set up yet.