Fix ChatColorPicker not initializing the Example Text at startup
The Example text wasn't getting set to the saved values in the profile
because SetColour() doesn't throw the event that we've been catching to
update the Example Text.  Now it does throw that event.

Also Profile now catches that event also so changes will get down into
ProfileData and get saved and IsModified() etc etc

This fixes #164.