Edit Default Profile experiment - test2
This second test release changes up some of the UI and underlying logic
on when / how to show it, pertaining to the edit-default-profile mode.

Along for the ride are a couple of innards changes and a couple of
bugfixes and feature twiddles for PowerBinder that were surfaced while
doing this work.

This is still pretty raw, and all input is welcome.