New "Edit Default Profile" functionality
This release adds a new "Edit Default Profile" entry to the "Profile"
menu that allows the user to open up the default profile for editing and
updating.  Prior to this, the only way to modify the default profile was
to create a new profile (which was automatically based on the default
profile), edit it, and then "Save as Default" which was fiddly and
unintuitive.

When editing the default profile, several bits of the UI are disabled,
most notably "Write Binds" and the normal Profile > Save menu items.
Profile > Save as Default still works, and attempting to close an edited
Default Profile will ask if you want to save and will do the right thing
if so.

There were also a number of under-the-hood cleanups, as well as a few
fixes to unrelated PowerBinder command bugs that were discovered when
testing the under-the-hood cleanups.

Check it out, report bugs.