Precache power icons at startup time
This spawns a thread that fetches Primary, Secondary, Epic, Pool,
Incarnate, and Misc power icons in the background, putting them in the
memory cache.

This eliminates UI lag when running from one of the binary releases that
uses the icon ZIP file and, for instance, clicking on a PowerPicker
which then needs to populate its entire menu / submenu structure.

Subjectively, this resulted in a very noticeable reduction in delay when
opening the first PowerPicker, on a binary prerelease version built from
this code, down to nearly instantaneous from over two seconds.

This doesn't seem to have any downside.  We shall see.