Another stab at this.  More below.
This one puts the toggle binds into reset and X000000.txt, as well as a
stanza that turns off all SoD powers.

This almost worked except the movement keys turned the SoD powers back
on and off.

So then I added "rebind movement keys to their basic forms" into the
X000000.txt file, which helped that problem, except re-activating SoD
didn't rebind the keys to dtrt.

This may be a complete rabbit hole.  Going to fall back to the previous
effort for a little bit and see if I can make it work.  Seems like it
was the right solution if the implementation could get ironed out.