Test release:  Major Movement Power Overhaul
This very early test release is a proof-of-concept of some major changes
to how Movement Powers are handled by BindControl.

Until now, BindControl would handle Speed on Demand binds for the various
travel power pools, and not much else.  This release adds quite a bit of
flexibility:

* Speed on Demand works as always, turning on movement powers while
  moving and turning them off, optionally switching to a lower-endurance
  defensive power, when stationary.
* The Primary and Defensive Powers are both selectable, in the
  somewhat-rare case that a Profile has more than one of them.
* Speed on Demand works via "Modes" -- you're in "Speed Mode" or "Fly
  Mode" etc, determining which travel powers will be activated by the
  movement keys.  Modes can be switched via keybinds, and the entire SoD
  system can be toggle on or off with a keybind.
* Power Toggle can toggle between the Primary Travel Power and the
  Defensive Power, or toggle either of those on and off.

Changed behaviors:
* Speed on Demand Modes now try to do a better job of shutting off
  travel powers from other Modes at switch-time.  Previously it was
  pretty easy to get in a state where more than one travel power was
  active at a time.
* Power Toggles will also turn off all other powers and disable Speed on
  Demand behavior until it's explicitly turned back on.
* Sprint is now a first-class Speed-on-Demand and Toggle power, although
  it is available to all characters.  The actual Sprint power is
  choosable, as always.

This release is *VERY* alpha-quality.  It may act weirdly when
traveling.  You will likely need to make liberal use of the Reset Binds
key (<kbd>CTRL</kbd>+<kbd>R</kbd> by default) to get things back to a somewhat
normal state.

Your profiles, if saved with this new version, will probably lose some
travel-power configuration if you take them back to a previous version
of BindControl.  Some things changed in an incompatible way, and while
old profiles should work with the new code, it's not possible to support
that the other way.  Non-travel-power configuration will be unaffected.
Make copies and backups if you have concerns.

Please check it out, and report all bugs and observations (and there
will be plenty of both) in the Homecoming Forums thread or as Github
issues:

https://forums.homecomingservers.com/topic/38674-bindcontrol-keybinds-and-popmenus-editor-and-manager/

https://github.com/emersonrp/bindcontrol/issues